Updated PR inline with discussed requirements
1) Changed string per HA naming conventions
2) Moved common code to module to reduce duplication
3) Removed apply_custom_configuration
4) Replaced individual quirk files with common file for all modules
